Move CellController to its own file
